Add tests for callback_shortcuts.0.dart API examples.
